+++ plugin-dev/Source/Sentry/Private/Tests/SentryUser.spec.cpp
@@ -23,7 +23,7 @@ void SentryUserSpec::Define()
 		It("should persist their values", [this]()
 		{
 			const FString TestUsername = FString(TEXT("SentryUser"));
-			const FString TestEmail = FString(TEXT("user-mail@sentry.com"));
+			const FString TestEmail = FString(TEXT("user-mail@example.org"));
 			const FString TestIpAddress = FString(TEXT("127.0.0.1"));
 			const FString TestId = FString(TEXT("123"));