+++ CHANGELOG.md
@@ -14,6 +14,9 @@
 - Bump Java SDK (Android) from v7.15.0 to v7.16.0 ([#665](https://github.com/getsentry/sentry-unreal/pull/665))
   - [changelog](https://github.com/getsentry/sentry-java/blob/main/CHANGELOG.md#7160)
   - [diff](https://github.com/getsentry/sentry-java/compare/7.15.0...7.16.0)
+- Bump Native SDK from v0.7.6 to v0.7.11 ([#667](https://github.com/getsentry/sentry-unreal/pull/667))
+  - [changelog](https://github.com/getsentry/sentry-native/blob/master/CHANGELOG.md#0711)
+  - [diff](https://github.com/getsentry/sentry-native/compare/0.7.6...0.7.11)
 
 ## 0.20.1