+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features
 
 - Assertions now print their callstack a log file ([#637](https://github.com/getsentry/sentry-unreal/pull/637))
+- Add LinuxArm64 support for UE plugin ([#672](https://github.com/getsentry/sentry-unreal/pull/672))
 
 ### Dependencies